requirements.txt
Brotli==1.0.9
click==8.0.4
contextvars==2.4
cycler==0.11.0
dash==2.6.2
dash-core-components==2.0.0
dash-extensions==0.0.71
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses==0.8
decorator==4.4.2
EditorConfig==0.12.3
Flask==2.0.3
Flask-Caching==1.10.1
Flask-Compress==1.13
glob2==0.7
imageio==2.15.0
immutables==0.19
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.1
jsbeautifier==1.14.7
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.4
more-itertools==8.14.0
networkx==2.5.1
numpy==1.19.5
opencv-python==4.6.0.66
pandas==1.1.5
Pillow==8.4.0
plotly==5.11.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
PyWavelets==1.1.1
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
tenacity==8.1.0
threadpoolctl==3.1.0
tifffile==2020.9.3
typing_extensions==4.1.1
Werkzeug==2.0.3
zipp==3.6.0
